msm_fb: Fix framebuffer console for buzz and click
Issue was quite obvious when this kernel was used in clockworkmod
recovery. The screen kept shifting and jumping around when navigating.
This issue was not present in the .32 kernel, so adapt code from there
to avoid using 32 pixel aligned buffers, which our devices doesn't seem
to support. Seems to be primarily for devices having an Adrendo GPU.